+++ DownloaderForReddit/gui/settings/core_settings_widget.py
@@ -19,6 +19,7 @@ class CoreSettingsWidget(AbstractSettingsWidget, Ui_CoreSettingsWidget):
         self.size_map = {x: 1024**count for count, x in enumerate(['Bytes', 'KB', 'MB', 'GB'])}
         for key, value in self.size_map.items():
             self.threshold_size_combo.addItem(key, value)
+            self.chunk_size_combo.addItem(key, value)
         self.select_user_base_directory_button.clicked.connect(
             lambda: self.select_directory_path(self.user_save_dir_line_edit))
         self.select_subreddit_base_directory_button.clicked.connect(
@@ -42,14 +43,19 @@ class CoreSettingsWidget(AbstractSettingsWidget, Ui_CoreSettingsWidget):
         self.finish_incomplete_downloads_checkbox.setChecked(
             self.settings.finish_incomplete_downloads_at_session_start)
         self.download_reddit_hosted_videos_checkbox.setChecked(self.settings.download_reddit_hosted_videos)
-        self.set_threshold_size(self.settings.multi_part_threshold)
+        self.multi_part_download_groupbox.setChecked(self.settings.use_multi_part_downloader)
+        self.set_size_options(self.settings.multi_part_threshold, self.threshold_size_combo,
+                              self.multipart_threshold_spinbox)
+        self.set_size_options(self.settings.multi_part_chunk_size, self.chunk_size_combo,
+                              self.multi_part_chunk_size_spinbox)
+        self.multi_part_thread_count_spinbox.setValue(self.settings.multi_part_thread_count)
 
-    def set_threshold_size(self, size):
+    def set_size_options(self, size, combo, spinbox):
         for key, value in sorted(self.size_map.items(), key=lambda x: x[1], reverse=True):
             if round(size / value, 2) > 0:
-                self.threshold_size_combo.setCurrentText(key)
+                combo.setCurrentText(key)
                 break
-        self.multipart_threshold_spinbox.setValue(round(size / value, 2))
+        spinbox.setValue(round(size / value, 2))
 
     def apply_settings(self):
         self.settings.user_save_directory = self.user_save_dir_line_edit.text()
@@ -65,9 +71,11 @@ class CoreSettingsWidget(AbstractSettingsWidget, Ui_CoreSettingsWidget):
         self.settings.finish_incomplete_downloads_at_session_start = \
             self.finish_incomplete_downloads_checkbox.isChecked()
         self.settings.download_reddit_hosted_videos = self.download_reddit_hosted_videos_checkbox.isChecked()
+        self.settings.use_multi_part_downloader = self.multi_part_download_groupbox.isChecked()
         threshold_size = \
             int(self.multipart_threshold_spinbox.value() * self.threshold_size_combo.currentData(Qt.UserRole))
         self.settings.multi_part_threshold = threshold_size
+        self.settings.multi_part_thread_count = self.multi_part_thread_count_spinbox.value()
 
     def select_directory_path(self, line_edit):
         text = line_edit.text()