+++ CHANGE_LOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## v3.2.2-beta
+
+### Minor Changes
+* Users, subreddits, posts, content, and comments can now be permanently deleted from the database along with any
+associated files.
+* Json imports now support legacy export files from previous versions.
+* More aspects of multi-part download option is now user controllable.
+* Improve user/subreddit date limit controls
+
+### Bug Fixes
+* Make multi-part downloader stoppable with rest of download cycle.
+* Fix content not downloading correctly when the author's account has been deleted or suspended.
+* Fix issue that caused application lock due to malfunctioning multi-part download.
+* Fix issues with extracting and downloading reddit videos
+
+
+
 ## v3.2.1-beta
 
 ### Bug Fixes